services/qwen-vl-ocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
transformers==4.44.2
accelerate==0.34.2
torch==2.3.1
huggingface_hub==0.24.6
